pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-dev-pipeline"
version = "3.0.0"
description = "通用 AI 自动化开发流水线平台 — 文档驱动任务拆解、动态机器池、零项目绑定"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "lucas-realman" },
]

dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "psutil>=5.9.0",
    "packaging>=21.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-json-report>=1.5",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
autodev = "orchestrator.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "smoke: L1 冒烟测试 — 每次 commit 执行",
    "component: L2 组件测试 — 每个模块核心逻辑",
    "integration: L3 集成测试 — 模块间端到端链路",
    "acceptance: L4 验收测试 — 里程碑人工+自动验收",
]
